Changes to code structure
