+++ public_holidays.gemspec
@@ -11,8 +11,8 @@ Gem::Specification.new do |spec|
 
   spec.summary       = %q{This gem provides the public holidays of 42 countries.}
   spec.description   = %q{This gem provides the public holidays of 42 countries in the world between the dates 01-01-2016 to 31-12-2016.}
-  spec.homepage      = "htto;//github.com/neosepulveda/public-holidays"
-  spec.lincenses      = ['MIT']
+  spec.homepage      = "http://github.com/neosepulveda/public-holidays"
+  spec.licenses      = ['MIT']
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
   # to allow pushing to a single host or delete this section to allow pushing to any host.
